requirements.txt
torch
torchvision
transformers
pillow
matplotlib
streamlit
pytesseract
pandas
